AdsDevice: use size_t for length parameter
The natural datatype for buffer sizes is size_t. Often we pass
sizeof(something) or some_container.length() to our ADS wrapper
functions Read/WriteReqEx2().
For ADS the on wire format for these length fields is only 32-Bit, so
we should verify that our length fits into these bits.
Code should call our AdsDevice wrapper instead of the plain C API
functions. So by changing the interface for AdsDevice to accept size_t,
we can drastically reduce the number of length <= uint32_t::max()
checks. We will introduce these checks in a follow up commit.

+++ AdsLib/AdsDevice.cpp
@@ -175,16 +175,16 @@ AdsHandle AdsDevice::OpenFile(const std::string& filename, const uint32_t flags)
     return {new uint32_t {handle}, {std::bind(&AdsDevice::CloseFile, this, std::placeholders::_1)}};
 }
 
-long AdsDevice::ReadReqEx2(uint32_t group, uint32_t offset, uint32_t length, void* buffer, uint32_t* bytesRead) const
+long AdsDevice::ReadReqEx2(uint32_t group, uint32_t offset, size_t length, void* buffer, uint32_t* bytesRead) const
 {
     return AdsSyncReadReqEx2(*m_LocalPort, &m_Addr, group, offset, length, buffer, bytesRead);
 }
 
 long AdsDevice::ReadWriteReqEx2(uint32_t    indexGroup,
                                 uint32_t    indexOffset,
-                                uint32_t    readLength,
+                                size_t      readLength,
                                 void*       readData,
-                                uint32_t    writeLength,
+                                size_t      writeLength,
                                 const void* writeData,
                                 uint32_t*   bytesRead) const
 {
@@ -197,7 +197,7 @@ long AdsDevice::ReadWriteReqEx2(uint32_t    indexGroup,
                                   );
 }
 
-long AdsDevice::WriteReqEx(uint32_t group, uint32_t offset, uint32_t length, const void* buffer) const
+long AdsDevice::WriteReqEx(uint32_t group, uint32_t offset, size_t length, const void* buffer) const
 {
     return AdsSyncWriteReqEx(GetLocalPort(), &m_Addr, group, offset, length, buffer);
 }
